order_microservice/requirements.txt
aiokafka==0.12.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==5.0.1
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.112.2
fastapi-cli==0.0.5
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
ifaddr==0.2.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kafka-python-ng==2.2.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
openpyxl==3.1.5
orjson==3.10.7
packaging==24.2
pika==1.3.2
pillow==10.4.0
pluggy==1.5.0
psycopg2-binary==2.9.9
py_eureka_client==0.11.13
pyasn1==0.6.0
pycparser==2.22
pydantic==2.10.2
pydantic-extra-types==2.9.0
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.2
reportlab==4.2.2
requests==2.32.3
rich==13.8.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlmodel==0.0.22
starlette==0.38.2
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.0.1
